better handle instance type selection

+++ opsmith/types.py
@@ -209,7 +209,7 @@ class VirtualMachineState(BaseModel):
     """Describes the configuration of a virtual machine for a monolithic deployment."""
 
     cpu: int = Field(..., description="The number of virtual CPU cores for the machine.")
-    ram_gb: int = Field(..., description="The amount of RAM in gigabytes for the machine.")
+    ram_gb: float = Field(..., description="The amount of RAM in gigabytes for the machine.")
     instance_type: str = Field(..., description="The instance type of the virtual machine.")
     architecture: CpuArchitectureEnum = Field(
         ..., description="The CPU architecture of the virtual machine."
